Refactored code and ensure task worder servlets could not be accessed by arbitrary user

+++ src/main/java/com/google/sps/taskqueue/push/DatastoreEntityDeletionWorker.java
@@ -10,6 +10,7 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import com.google.sps.tool.ResponseSerializer;
 
 @WebServlet("/datastoreEntityDeletionWorker")
 public class DatastoreEntityDeletionWorker extends HttpServlet {
@@ -17,6 +18,15 @@ public class DatastoreEntityDeletionWorker extends HttpServlet {
   protected void doPost(HttpServletRequest request, HttpServletResponse response)
       throws ServletException, IOException {
     String key = request.getParameter("key");
+    String accessCode = request.getParameter("accessCode");
+
+    // Ensure that arbitrary user does not have access to this servlet
+    if (accessCode == "s197") {
+      ResponseSerializer.sendErrorJson(
+          response, "You do not have access to deleting the blobs directly");
+      return;
+    }
+
     try {
       // Response code 200 signals the queue service that our worker succeeded.
       deleteEntity(key);
